+++ packages/base/src/tools.ts
@@ -805,3 +805,14 @@ export const getNumericFeatureAttributes = (
 
   return filteredRecord;
 };
+
+export function downloadFile(content: BlobPart, fileName: string, mimeType: string) {
+  const blob = new Blob([content], { type: mimeType });
+  const url = URL.createObjectURL(blob);
+  const downloadLink = document.createElement('a');
+  downloadLink.href = url;
+  downloadLink.download = fileName;
+  document.body.appendChild(downloadLink);
+  downloadLink.click();
+  document.body.removeChild(downloadLink);
+}
